Data on member states

This dataset includes data on European Union (EU) member states. There is one observation per member state. The dataset includes information about accession, the political system of the member state, participation in the European Economic and Monetary Union (EMU) and the Schengen Area, legal obligations and opt-outs, and membership in other international organizations.

member_states

Format

A data frame with 58 variables:

key_id

Numeric. An ID number that uniquely identifies each observation.

member_state_id

Numeric. An ID number that uniquely identifies each member state.

member_state

String. The name of the member sate.

member_state_code

String. A two-letter code assigned to each member state by the Commission.

iso_alpha_2_code

String. The ISO 2-letter code for the member state.

iso_alpha_3_code

String. The ISO 3-letter code for the member state.

languages

String. The official languages of the member state. If a member states has more than one official language, they are listed, separated by a comma.

accession_date

Date. The date that the member state acceded to the EU in the format YYYY-MM-DD.

accession_year

Numeric. The year that the member state acceded to the EU.

current_member

Dummy. A dummy variable indicating whether the member state is currently a member of the EU. Coded 1 if the member state is currently a member and 0 otherwise.

exit_date

Date. If the member state is no longer a member of the EU, the date that the member state exited the EU in the format YYYY-MM-DD. Coded NA if not applicable.

exit_year

Numeric. If the member state is no longer a member of the EU, the year that the member state exited to the EU. Coded NA if not applicable.

wave_1

Dummy. A dummy variable indicating whether the member state was a founding member of the EU. Coded 1 if the member state was a founding member and 0 otherwise.

wave_2

Dummy. A dummy variable indicating whether the member state was part of the second accession wave (1973). Coded 1 if the member state was part of the second wave and 0 otherwise.

wave_3

Dummy. A dummy variable indicating whether the member state was part of the third accession wave (1981). Coded 1 if the member state was part of the third wave and 0 otherwise.

wave_4

Dummy. A dummy variable indicating whether the member state was part of the fourth accession wave (1986). Coded 1 if the member state was part of the fourth wave and 0 otherwise.

wave_5

Dummy. A dummy variable indicating whether the member state was part of the fifth accession wave (1995). Coded 1 if the member state was part of the fifth wave and 0 otherwise.

wave_6

Dummy. A dummy variable indicating whether the member state was part of the sixth accession wave (2004). Coded 1 if the member state was part of the sixth wave and 0 otherwise.

wave_7

Dummy. A dummy variable indicating whether the member state was part of the seventh accession wave (2007). Coded 1 if the member state was part of the seventh wave and 0 otherwise.

wave_8

Dummy. A dummy variable indicating whether the member state was part of the eighth accession wave (2013). Coded 1 if the member state was part of the eighth wave and 0 otherwise.

eu_15

Dummy. A dummy variable indicating whether the member state is part of the EU15, which refers to the members that joined prior to the 2004 expansion into Central and Eastern Europe. Coded 1 if the member state is part of the EU15 and 0 otherwise.

eu_19

Dummy. A dummy variable indicating whether the member state is part of the EU19, which refers to the members of the Eurozone. Coded 1 if the member state is part of the EU19 and 0 otherwise.

eu_10

Dummy. A dummy variable indicating whether the member state is part of the EU10, which refers to the members that joined as part of the 2004 expansion into Central and Eastern Europe (including Cyprus and Malta). Coded 1 if the member state is part of the EU10 and 0 otherwise.

eu_11

Dummy. A dummy variable indicating whether the member state is part of the EU11, which refers to the Central and Eastern European member states that joined between 2004 and 2013 (excluding Cyprus and Malta). Coded 1 if the member state is part of the EU11 and 0 otherwise.

eu_27_old

Dummy. A dummy variable indicating whether the member state is part of the EU27 under the original meaning, which refers to all member states except Croatia. Coded 1 if the member state is part of the EU27 and 0 otherwise.

eu_27_new

Dummy. A dummy variable indicating whether the member state is part of the EU27 under the new meaning, which refers to all member states except the United Kingdom. Coded 1 if the member state is part of the EU27 and 0 otherwise.

eu_28

Dummy. A dummy variable indicating whether the member state is part of the EU28, which refers to all member states. Coded 1 if the member state is part of the EU28 and 0 otherwise.

political_system

String. The political system of the member state. Possible values include, in decreasing order of frequency: unitary parliamentary republic, unitary parliamentary constitutional monarchy, unitary semi-presidential republic, federal parliamentary republic, federal parliamentary constitutional monarchy, and unitary presidential republic.

federal

Dummy. A dummy variable indicating whether the member state has a federal system. Coded 1 if the member state has a federal system and 0 otherwise.

unitary

Dummy. A dummy variable indicating whether the member state has a unitary system. Coded 1 if the member state has a unitary system and 0 otherwise.

parliamentary

Dummy. A dummy variable indicating whether the member state has a parliamentary system. Coded 1 if the member state has a parliamentary system and 0 otherwise.

presidential

Dummy. A dummy variable indicating whether the member state has a presidential system. Coded 1 if the member state has a presidential system and 0 otherwise.

semi_presidential

Dummy. A dummy variable indicating whether the member state has a semi-presidential system. Coded 1 if the member state has a semi-presidential system and 0 otherwise.

republic

Dummy. A dummy variable indicating whether the member state is a republic. Coded 1 if the member state is a republic and 0 otherwise.

constitutional_monarchy

Dummy. A dummy variable indicating whether the member state is a constitutional monarchy. Coded 1 if the member state is a constitutional monarchy and 0 otherwise.

emu

Dummy. A dummy variable indicating whether the member state is a member of the European Economic and Monetary Union (EMU). Coded 1 if the member state is a member of the EMU and 0 otherwise.

emu_accession_date

Date. If the member state is a member of the EMU, the date that the member state acceded to the EMU in the format YYYY-MM-DD. Coded NA if not applicable.

emu_accession_year

Year. If the member state is a member of the EMU, the year that the member state acceded to the EMU. Coded NA if not applicable.

currency

String. The name of the currency used by the member state.

pre_emu_currency

String. The name of the currency used by the member state before the creation of the EMU.

schengen_area

Dummy. A dummy variable indicating whether the member state is a member of the Schengen Area. Coded 1 if the member state is a member of the Schengen Area and 0 otherwise.

schengen_date_signed

Date. If the member state is a member of the Schengen Area, the date that the member state signed the Schengen Agreement in the format YYYY-MM-DD. Coded NA if not applicable.

schengen_year_signed

Year. If the member state is a member of the Schengen Area, the year that the member state signed the Schengen Agreement. Coded NA if not applicable.

schengen_date_implemented

Date. If the member state is a member of the Schengen Area, the date that the member state implemented the Schengen Agreement in the format YYYY-MM-DD. Coded NA if not applicable.

schengen_year_implemented

Year. If the member state is a member of the Schengen Area, the year that the member state implemented the Schengen Agreement. Coded NA if not applicable.

emu_obligated

Dummy. A dummy variable indicating whether the member state is legally obligated to join in the EMU in the future. Coded 1 if the member state has an obligation to join and 0 otherwise.

schengen_obligated

Dummy. A dummy variable indicating whether the member state is legally obligated to join in the Schengen Area in the future. Coded 1 if the member state has an obligation to join and 0 otherwise.

emu_opt_out

Dummy. A dummy variable indicating whether the member state has a legal opt-out for the EMU. Coded 1 of the member state has an opt-out and 0 otherwise.

schengen_opt_out

Dummy. A dummy variable indicating whether the member state has a legal opt-out for the Schengen Area. Coded 1 of the member state has an opt-out and 0 otherwise.

csdp_opt_out

Dummy. A dummy variable indicating whether the member state has a legal opt-out for the Common Security and Defense Policy (CSDP). Coded 1 of the member state has an opt-out and 0 otherwise.

cfr_opt_out

Dummy. A dummy variable indicating whether the member state has a legal opt-out for the Charter of Fundamental Rights (CFR). Coded 1 of the member state has an opt-out and 0 otherwise.

afsj_opt_out

Dummy. A dummy variable indicating whether the member state has a legal opt-out for the Area of Freedom, Security and Justice (AFSJ). Coded 1 of the member state has an opt-out and 0 otherwise.

nato

Dummy. A dummy variable indicating whether the member state is also a member of the North Atlantic Treaty Organization (NATO). Coded 1 if the member state is a member and 0 otherwise.

oecd

Dummy. A dummy variable indicating whether the member state is also a member of the Organization for Economic Cooperation and Development (OECD). Coded 1 if the member state is a member and 0 otherwise.

benelux_union

Dummy. A dummy variable indicating whether the member state is also a member of the Benelux Union. Coded 1 if the member state is a member and 0 otherwise.

nordic_council

Dummy. A dummy variable indicating whether the member state is also a member of the Nordic Council. Coded 1 if the member state is a member and 0 otherwise.

baltic_assembly

Dummy. A dummy variable indicating whether the member state is also a member of the Baltic Assembly. Coded 1 if the member state is a member and 0 otherwise.

visegrad_group

Dummy. A dummy variable indicating whether the member state is also a member of the Visegrad Group. Coded 1 if the member state is a member and 0 otherwise.